admin服务新增功能：（1）获取待处理板块申请记录数量；（2）获取待处理板主申请数量；（3）获取待处理举报帖子数量；（4）获取待处理举报评论数量

+++ thinktank-generator/src/main/java/com/thinktank/generator/vo/UserStatisticsVo.java
@@ -15,7 +15,7 @@ public class UserStatisticsVo {
     private Integer userCount;
 
     @ApiModelProperty("新用户数量")
-    private Long newUserCount;
+    private Integer newUserCount;
 
     @ApiModelProperty("在线人数")
     private Integer onlineCount;
